[fix] let backend handle petition field errors (#1576)
Se espera a que el backend devuelva el error antes de mostrar el dialogo
de conflictos por referencias en condiciones de otros campos.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1205077061076498

+++ server/src/graphql/petition/mutations/core.ts
@@ -999,11 +999,6 @@ export const deletePetitionField = mutationField("deletePetitionField", {
     force: booleanArg({ default: false }),
   },
   resolve: async (_, args, ctx) => {
-    const replies = await ctx.petitions.loadRepliesForField(args.fieldId);
-    if (!args.force && replies.length > 0) {
-      throw new ApolloError("The petition field has replies.", "FIELD_HAS_REPLIES_ERROR");
-    }
-
     const petitionFields = await ctx.petitions.loadFieldsForPetition.raw(args.petitionId);
     if (
       petitionFields.some(
@@ -1014,6 +1009,11 @@ export const deletePetitionField = mutationField("deletePetitionField", {
         "The petition field is being referenced in another field.",
         "FIELD_IS_REFERENCED_ERROR",
       );
+    }
+
+    const replies = await ctx.petitions.loadRepliesForField(args.fieldId);
+    if (!args.force && replies.length > 0) {
+      throw new ApolloError("The petition field has replies.", "FIELD_HAS_REPLIES_ERROR");
     }
 
     return await ctx.petitions.deletePetitionField(args.petitionId, args.fieldId, ctx.user!);
